fix tests, stub VRScene.setupRenderer for tests for travis CI (#195)
